packages/dhis2w-browser/pyproject.toml
[project]
name = "dhis2w-browser"
version = "0.5.0"
description = "Playwright-based helpers for DHIS2 UI automation."
readme = "README.md"
authors = [{ name = "Morten Hansen", email = "morten@winterop.com" }]
requires-python = ">=3.13"
dependencies = [
  "dhis2w-client>=0.5.0,<0.6",
  "pillow>=12.2.0",
  "playwright>=1.58",
  "pydantic>=2.13",
]

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"
